feat: finish challenge
Finish readme.md file description.
menu.sh not work because it isnt be able to pass arg to menu scripts

+++ bin/deploy-locally-image.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
-docker build -t wefox.challenge.app:5000/http-echo:v1 .
-docker push wefox.challenge.app:5000/http-echo:v1
+if [ -z "$1"]; then
+    echo "image version is required, example: ./deploy-remote-image.sh 0.2.1"
+    exit
+fi
+docker build -t wefox.challenge.app:5000/http-echo:$1 .
+docker push wefox.challenge.app:5000/http-echo:$1
 kubectl delete -f ./config/wefox-challenge-conf-local.yml
 kubectl delete -f ./config/wefox-challenge-conf.yml
+sed -i "s/image-version/$1/g" ./config/wefox-challenge-conf-local.yml
 kubectl crete -f ./config/wefox-challenge-conf-local.yml
 sleep 1
 kubectl get pods
