+++ bin/deploy-remote-image.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
+if [ -z "$1"]; then
+    echo "image version is required, example: ./deploy-remote-image.sh 0.2.1"
+    exit
+fi
 kubectl delete -f ./config/wefox-challenge-conf.yml
 kubectl delete -f ./config/wefox-challenge-conf-local.yml
+sed -i "s/image-version/$1/g" ./config/wefox-challenge-conf.yml
 kubectl create -f ./config/wefox-challenge-conf.yml
 sleep 4
 kubectl get pods
